// src/app/model/comprovante-residencia.model.ts
export class ComprovanteResidencia {
  cep: string;
  dataEmissao: string;
  nome: string;
  emissor: string;
  dataReferencia: string;
  endereco: string;
  uf: string;
  municipio: string;
  localidade: string;
  tipoLogradouro: string;
  bairro: string;
  complemento: string;
  estado: string;
  numero: string;
  mesAnoReferencia: string;
  cpf: string;
  tipoEndereco: string;
  finalidade: string;
  correspondecia: string;
  comprovacao: string;
}